refactor(core): organize wallet domain
