+++ core/src/services/wallet/WalletService.ts
@@ -1,8 +1,8 @@
 import { ClientSession } from 'mongoose';
-import UserWallet from '../models/UserWallet';
-import Transaction, { type ITransaction } from '../models/Transaction';
-import { getUserConnection } from '../config/db';
-import { AppError } from '../utils/AppError';
+import UserWallet from '../../models/UserWallet';
+import Transaction, { type ITransaction } from '../../models/Transaction';
+import { getUserConnection } from '../../config/db';
+import { AppError } from '../../utils/AppError';
 import { getPrimaryPlanCreditCount } from "@esparex/shared";
